Revert refactoring
