+++ src/components/property-tree/property-tree.tsx
@@ -26,8 +26,7 @@ export const PropertyTree = ({ asset, childAssets }: PropertyTreeProps): JSX.Ele
   const [childNodes, setChildNodes] = useState<TreeAsset[]>([]);
 
   useEffect(() => {
-    const childrenAssets = addChildrenAssets(childAssets, asset.id);
-    setChildNodes(childrenAssets);
+    setChildNodes(addChildrenAssets(childAssets, asset.id));
   }, [childAssets, asset]);
 
   useEffect(() => {
