+++ src/components/related-assets/related-assets.tsx
@@ -1,65 +1,58 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
+import { Link as RouterLink } from "react-router-dom";
 
-import { RelatedAssetGroup } from "./related-asset-group";
-import { RelatedAsset } from "./types";
-import { getAllRelatedAssets, organiseRelatedAssetsByType } from "./utils";
+import { Link, LinkBox, LinkOverlay } from "@mtfh/common/lib/components";
+import { SearchCard } from "@mtfh/search";
 
-import { Asset, ParentAsset } from "@mtfh/common/lib/api/asset/v1";
-import { Center, Spinner } from "@mtfh/common/lib/components";
+import { RelatedAsset as IRelatedAsset } from "views/related-assets-view/utils";
 
-interface RelatedAssetsProps {
-  parentAssets: ParentAsset[];
-  childrenAssets: Asset[] | undefined;
-  loading: boolean;
+export interface RelatedAssetsProps {
+  assetType: string;
+  relatedAssets: IRelatedAsset[];
 }
 
-export const RelatedAssets = (props: RelatedAssetsProps) => {
-  const { parentAssets, childrenAssets, loading } = props;
+export const RelatedAssets = ({ assetType, relatedAssets }: RelatedAssetsProps) => {
+  const getAssetTypeHeading = () => {
+    return assetType.charAt(assetType.length - 1).toLowerCase() === "s"
+      ? assetType
+      : `${assetType}s`;
+  };
 
-  const [relatedAssets, setRelatedAssets] = useState<RelatedAsset[]>([]);
-  const [relatedAssetsByType, setRelatedAssetsByType] = useState<any>(undefined);
-
-  useEffect(() => {
-    if (parentAssets?.length && childrenAssets) {
-      setRelatedAssets(getAllRelatedAssets(parentAssets, childrenAssets));
-    }
-  }, [parentAssets, childrenAssets]);
-
-  useEffect(() => {
-    if (relatedAssets.length) {
-      const assetsByType = organiseRelatedAssetsByType(relatedAssets);
-
-      setRelatedAssetsByType(assetsByType);
-    }
-  }, [relatedAssets]);
-
-  if (loading) {
-    return (
-      <Center>
-        <Spinner />
-      </Center>
-    );
-  }
-
-  if (relatedAssets.length || !relatedAssetsByType) {
-    return (
-      <p className="lbh-body-m" data-testid="no-related-assets-message">
-        There are no related assets for this property.
-      </p>
-    );
-  }
+  const renderRelatedAssets = () => {
+    return relatedAssets.map((relatedAsset) => {
+      return <RelatedAsset relatedAsset={relatedAsset} key={relatedAsset.id} />;
+    });
+  };
 
   return (
-    <>
-      {Object.keys(relatedAssetsByType).map((assetType) => (
-        <React.Fragment key={assetType}>
-          <RelatedAssetGroup
-            assetType={assetType}
-            relatedAssets={relatedAssetsByType[assetType]}
-          />
-          <hr style={{ borderTop: "1px solid #e7eaec" }} />
-        </React.Fragment>
-      ))}
-    </>
+    <section>
+      <p className="lbh-body-m">{getAssetTypeHeading()}</p>
+      <div className="mtfh-card-list" data-testid="related-asset-items-group">
+        {renderRelatedAssets()}
+      </div>
+    </section>
+  );
+};
+
+export interface RelatedAssetProps {
+  relatedAsset: IRelatedAsset;
+}
+
+export const RelatedAsset = ({ relatedAsset }: RelatedAssetProps) => {
+  return (
+    <LinkBox data-testid="related-asset">
+      <SearchCard>
+        <LinkOverlay>
+          <Link
+            className="mtfh-search-tenure__title"
+            as={RouterLink}
+            to={`/property/${relatedAsset.id}`}
+            variant="text-colour"
+          >
+            {relatedAsset.name}
+          </Link>
+        </LinkOverlay>
+      </SearchCard>
+    </LinkBox>
   );
 };
