--- src/components/related-assets/types.ts
@@ -1,5 +1,0 @@
-export interface RelatedAsset {
-  id: string;
-  name: string;
-  type: string;
-}
