+++ src/views/related-assets-view/layout.tsx
@@ -1,18 +1,82 @@
-import React, { ReactNode } from "react";
+import React, { useEffect, useState } from "react";
 import { Link as RouterLink } from "react-router-dom";
 
-import { Asset } from "@mtfh/common/lib/api/asset/v1";
-import { Link } from "@mtfh/common/lib/components";
+import { RelatedAssets } from "../../components/related-assets/related-assets";
+import { RelatedAsset, getAllRelatedAssets, organiseRelatedAssetsByType } from "./utils";
+
+import { Asset, ParentAsset } from "@mtfh/common/lib/api/asset/v1";
+import { Center, Link, Spinner } from "@mtfh/common/lib/components";
+
+// import React, { ReactNode } from "react";
+// import { Link as RouterLink } from "react-router-dom";
+
+// import { Asset } from "@mtfh/common/lib/api/asset/v1";
+// import { Link } from "@mtfh/common/lib/components";
 
 import "./styles.scss";
 
 interface RelatedAssetsLayoutProps {
   asset: Asset;
-  children: ReactNode;
+  parentAssets: ParentAsset[];
+  childrenAssets: Asset[] | undefined;
+  loading: boolean;
 }
 
-export const RelatedAssetsLayout = (props: RelatedAssetsLayoutProps): JSX.Element => {
-  const { asset, children } = props;
+export const RelatedAssetsLayout = ({
+  asset,
+  parentAssets,
+  childrenAssets,
+  loading,
+}: RelatedAssetsLayoutProps): JSX.Element => {
+  const [relatedAssets, setRelatedAssets] = useState<RelatedAsset[]>([]);
+  const [relatedAssetsByType, setRelatedAssetsByType] = useState<any>(undefined);
+
+  useEffect(() => {
+    if (parentAssets?.length || childrenAssets) {
+      setRelatedAssets(getAllRelatedAssets(parentAssets, childrenAssets));
+    }
+  }, [parentAssets, childrenAssets]);
+
+  useEffect(() => {
+    if (relatedAssets.length) {
+      const assetsByType = organiseRelatedAssetsByType(relatedAssets);
+      setRelatedAssetsByType(assetsByType);
+    }
+  }, [relatedAssets]);
+
+  const assetHasRelatedAssets = (): boolean => {
+    return !!relatedAssets.length;
+  };
+
+  const renderRelatedAssets = () => {
+    if (loading) {
+      return (
+        <Center>
+          <Spinner />
+        </Center>
+      );
+    }
+
+    // If we have related assets and the relatedAssetsByType object we render RelatedAssets components for each Asset type
+    if (assetHasRelatedAssets() && relatedAssetsByType) {
+      return Object.keys(relatedAssetsByType).map((assetType) => {
+        return (
+          <section key={assetType}>
+            <RelatedAssets
+              assetType={assetType}
+              relatedAssets={relatedAssetsByType[assetType]}
+            />
+            <hr style={{ borderTop: "1px solid #e7eaec" }} />
+          </section>
+        );
+      });
+    }
+    return (
+      <p className="lbh-body-m" data-testid="no-related-assets-message">
+        There are no related assets for this property.
+      </p>
+    );
+  };
 
   return (
     <>
@@ -30,7 +94,7 @@ export const RelatedAssetsLayout = (props: RelatedAssetsLayoutProps): JSX.Elemen
       </h2>
       <hr style={{ borderTop: "1px solid #e7eaec" }} />
 
-      {children}
+      {renderRelatedAssets()}
     </>
   );
 };
